Remove switchToModelVersionAt

+++ src/core/packages/saved-objects/migration-server-internal/src/actions/cleanup_unknown_and_excluded.ts
@@ -108,6 +108,9 @@ export const cleanupUnknownAndExcluded = ({
         },
       };
 
+      if (2 === 2) {
+        return TaskEither.right()
+      }
       return deleteByQuery({
         client,
         indexName,
